+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to `pfctl_exporter` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Parse `pfctl -vvs Tables` output into `pfctl_table_*` metrics.
+
+
 ## [v0.4.0] - 2023-12-18
 
 ### Added